penumbra_sdk_sct
Module component

Available on crate feature component only.

The Sct component contains the interface to the chain’s state commitment tree and nullifier set. It also serves as tracking the various chain clocks, whether logical, like an epoch index, or a block height, or physical, like block timestamps.

Re-exports

pub use sct::StateReadExt;

pub use sct::StateWriteExt;

Modules

clock

Blockchain clocks: epoch indices, block heights and timestamps.

rpc

Implementation of the SCT component query server.

sct

The SCT component implementation.

source

Tracking commitment sources within a block execution.

tree

Mediate access to the state commitment tree and related data.